Yolandah_Sibanda_H180401E/requirements.txt
Flask==1.1.1
nltk==3.4.5
Flask_WTF==0.14.3
WTForms==2.3.1
pandas==0.25.1
scikit_learn==0.23.1
gunicorn==20.0.4
joblib==0.14.1
newsapi-python==0.2.6